完成多源知识导入后端，解析文件为纯文本

+++ src/main/java/com/camplus/admin/Mappers/UserContributionMapper.java
@@ -20,10 +20,10 @@ public interface UserContributionMapper {
     @Update("UPDATE user_contributions SET status = #{status}, " +
             "review_comment = #{reviewComment} " +
             "WHERE contribution_id = #{contributionId}")
-    int updateReviewData(@Param("contributionId") Long contributionId,
+    int updateReviewData(@Param("contributionId") Integer contributionId,
                          @Param("status") Integer status,
                          @Param("reviewComment") String reviewComment);
 
     @Select("SELECT * FROM user_contributions WHERE contribution_id = #{contributionId}")
-    UserContribution selectById(@Param("contributionId") Long contributionId);
+    UserContribution selectById(@Param("contributionId") Integer contributionId);
 }
